#12 [move] : TodayTopData domain에 선언
